Multiple passes optimization Conditional jumps normalizer evaluates all literal expressions Proper handling of the END instruction in Data Flow optimizer Moved test resources to correct packages

+++ mindcode/src/main/java/info/teksol/mindcode/processor/IntVariable.java
@@ -38,7 +38,7 @@ public class IntVariable extends AbstractVariable {
         MindustryValueType type = var.getMindustryValueType();
         switch (type) {
             case NULL, OBJECT -> setObject(var.getObject());
-            default -> setDoubleValue(var.getDoubleValue());
+            default -> setIntValue(var.getIntValue());
         }
         setType(type);
     }
